WiP: Some fun with configuration and reaaranging components
