+++ src/main.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
-import './debug.css'
+import 'utils/development/debug.css'
 import App from './App.tsx'
 
 createRoot(document.getElementById('root')!).render(
